lambda/requirements.txt
boto3==1.34.20
botocore==1.34.20
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
jmespath==1.0.1
joblib==1.3.2
nltk==3.8.1
praw==7.7.1
prawcore==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2021.8.3
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
tqdm==4.66.1
update-checker==0.18.0
urllib3==1.26.18
websocket-client==1.7.0
